pyproject.toml
[tool.poetry]
name = "dau-issue"
version = "0.1.0"
description = ""
authors = [
    "Anton Lysenkov <soul4code@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.7.17"
django = "1.11.24"
psycopg2-binary = {version = "2.8.6", platform = "darwin"}
psycopg2 = {version = "2.8.6", platform = "linux"}
dotenv = "^0.9.9"
sqlalchemy = "^2.0.44"
django-extensions = "3.0.3"
celery = "5.2.7"
django-cte = "1.3.3"
importlib-metadata = "4.13.0"

[tool.poetry.group.dev.dependencies]
tqdm = "^4.67.1"
factory-boy = "3.2.1"
asyncpg = "0.28.0"
ipython = "7.34.0"
pytest = "7.4.4"
pytest-django = "4.5.2"
ruff = "^0.14.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "dau_issue.settings"
python_files = ["__tests__/test_*.py"]

[tool.ruff]
line-length = 100
exclude = [
    "*/migrations/*",
    "env/",
    "venv/",
    ".venv/",
    "manage.py",
]

[tool.ruff.lint]
preview = true
ignore = [
    "F821", # undefined name 'name'
    "E731", # Do not assign a lambda expression, use a def"
    "E721", # do not compare types, use isinstance()
]
select = [
    "F",
    "E",
    "W",
]

[tool.ruff.lint.per-file-ignores]
# https://docs.astral.sh/ruff/settings/#lint_per-file-ignores
"__init__.py" = [
    "F401",  # unused-import
]
"conftest.py" = [
    "F403",  # import *
]
"*/migrations/*" = [
    "E501",  # line-too-long
    "F401",  # unused-import
    "I001",  # unsorted-imports
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false
combine-as-imports = true
lines-after-imports = 2

[tool.ruff.format]
quote-style = "single"